rename testing compat module
